Determine model sources earlier on.
We need to do this when the models are initially read in,
so that their datasets get set up before we try to use them
in modeling protocols. This necessitates postponing the
calculation of source seq_id_{begin,end} until we actually need
them, since they may need the model seq_id range which isn't set
up until later.