Move protocol output code to its own class.
Rather than subclassing Representation, we now subclass
from a ProtocolOutput class. This class is solely responsible
for recording all details of a modeling protocol. By pulling
this code out of Representation, we allow for the possibility
of having multiple ProtocolOutput objects, and also for
attaching them to PMI 2's System class. The mmCIF module
provides a subclass of ProtocolOutput that records the protocol
in an mmCIF file.