Add support for 2DEM raw micrographs.
We can now annotate our 2DEM restraints with information on
where the raw micrographs used to create the class averages
are stored. While this isn't used in the modeling, it is
useful for people that want to rederive the averages.
The mmCIF writer will pick up this information and encode
it in the output mmCIF file.