Ignore crosslinks that refer to missing components.
Our input crosslink dataset may contain crosslinks to
components that we don't model. This means that we have
no data (entity ID, sequence) on those components to add
to the mmCIF file. For now, just ignore such crosslinks.
In future, we should probably add at least a dummy entity
so that we don't discard this data.