Allow adding components that are not modeled.
For example, the experimental information may have been collected
for a system larger than the one modeled. In order to correctly
represent the data, IMP needs to know about the extra components.
These are added to a separate mmCIF assembly, and are not assigned
any structure (thus, no asym/chain ID).