Improve support for nested rigid bodies.
If a rigid member is itself a rigid body, make sure we
set sensible coordinates for it from the RMF file rather
than just skipping it as we did before.
Note that we still don't properly handle the case where
such a rigid body itself contains members; this will now
throw an exception.