Don't run slow examples in debug mode.
These examples run quite quickly in fast or release
builds (no checks or only usage checks) but are very very
slow in debug mode (internal checks). Generally we try to
run examples as part of the test suite with debug builds
anyway, but these are just too slow and time out,
so instead we'll skip them for now.