Add restraint for torsion of planes formed by particles
When a chain of nearly identical rigid bodies should have a regular
twist, a triplet of particles corresponding to one another in the rigid
bodies may be used to restrain the dihedrals. This should really only
be used when the curvature of the chain is expected to be low, such
that the planes are more or less parallel.