Allow shuffle to prevent collision with fixed objects
When shuffling only part of the system, collision should be prevented
with other parts of the system that remain fixed.  These hierarchies or
PMI objects can be added to `hierarchies_included_in_collision` where
the particles of these components are added to the `all_idxs` set,
which is then used to cdetermine close pairs after a shuffle step.