Make TopologyReader.components private.
If the member is public then it is part of the API, should
be documented, and should not change without warning. Since
we can't guarantee any of these things, we should make it
private. As an example, it used to be called component_list
and some scripts (e.g. the PMI tutorial) used it, and so failed
when it changed name. Provide a (deprecated) component_list
property to support old scripts.