Generalize linking out to external files.
Update to use new mmCIF tables for linking to external
files. All such files are now listed in the ihm_external_files
table and are linked with a file_id from other tables. This
also allows use of multiple repositories in our modeling, as
well as files that live in no repository (local files on disk).