Remove ihm_starting_model_seq_dif.db_entity_id.
This field refers to the entity_id of the sequence in the
.cif file of the crystal structure of the starting subunit
model, not to an entity in our own mmCIF file. Since we don't
currently support reading crystal structures from .cif files,
the field is (now) non-mandatory, and the remaining fields
(db_asym_id, db_comp_id and db_seq_id) should uniquely define
the residue anyway, it's probably better for now not to emit
an entity_id. Closes ihmwg/IHM-dictionary#28.