Handle PDB files that are derived from something else.
For example, the modeling of the Mediator complex uses several
PDB files that are really official PDB structures, just rotated
and translated and with header information stripped. Allow such
files to be simply annotated with TITLE and EXPDTA records to
point back to the original PDB IDs or the DOI of a published
comparative model. Relates ihmwg/IHM-dictionary#29.