Add basic Travis setup.
It's not practical to build, test, and collect coverage of all
of IMP and PMI every time we make a change to PMI, but we can
compromise by testing only the Python parts of PMI (which is the
vast majority of the code) combined with a recent IMP binary
distribution. This should be fairly rapid and will let us know
if any of the (tested) Python code is broken.