Don't center models for mmCIF output by default.
For the sake of reproducing what was done in the modeling
as well as possible, it's better to use the same model
coordinates in the RMF and the mmCIF output (previously mmCIF
output was centered). This also ensures that trajectories
or ensembles are in the same reference frame as mmCIF and so work
correctly. Provide a Transformation3D for each model
so that this behavior can be overridden. (We do this for Nup84 since
the trajectories there are multimodel PDB files, which have already
been centered, so in this case it makes sense to center the mmCIF
output to match. Relates ihmwg/IHM-dictionary#30.)