Map chain IDs via component names.
We used to obtain the chain ID by using the Output class,
which simply assumes that the nth component uses the nth
letter in the alphabet as its chain ID. This won't necessarily
work with multiple states, since a later state may be missing
components that are present in an earlier state, and thus have
gaps in the chain IDs. Allow for this by mapping particles to
component names instead, which can then be unambiguously mapped
to chain IDs.