Assign our own chain IDs for mmCIF output.
When doing multi-state modeling the chain IDs returned by
the Output class may be incorrect (since they are sequential
but states may have gaps in the chain IDs). Since we have the
component name (which is unique) we can map these chain IDs
to the 'real' chain IDs, so that our mmCIF output is correct.