Clean up temporary RMF hierarchies
When we set coordinates from an existing RMF file,
we read in that RMF, create IMP hierarchies
corresponding to that file, and then copy information
from those hierarchies to the system of interest. This
leaks memory because the temporary hierarchies are never
destroyed. Fix this by explicitly destroying the hierarchies.