Support starting integrative models
Allow for some of the input structures we use in modeling
to be derived from earlier integrative models. These are
denoted with PDB headers similar to those for derived
comparative models (EXPDTA and TITLE lines) and will
populate the relevant mmCIF categories, linking back to
the paper (via DOI) for the original models.