Allow marking models as unrestrained
When outputing an mmCIF file, the code currently assumes
that every model is restrained by every restraint. This is
not the case for the FG repeat model used in our NPC modeling.
As a temporary hack, allow manually marking individual models as
unrestrained, so restraint information isn't output for them.