Handle all mutations in starting models
If the starting model sequence differs from that
of the output model, add a SeqDif object to the mmCIF
file and print a warning (as multiple mutations may be
an indication that the offset is incorrect). For example,
the Nup84 study mutates MSE to LEU in Sec13 (to reverse
the engineered mutation in the crystal structure). Don't warn
for the common MSE -> MET mutation.