Allow IHM and PMI numbering to differ
Residue numbering in IHM always starts at 1 and is
contiguous. PMI is usually the same, taking its numbering from
the input FASTA file, but can differ if the FASTA sequence
contains N-terminal gaps (which are ignored by IHM but are
counted by PMI). Track a PMI offset for all IHM Entities
so that we can get the IHM numbering correct. Populate the
mmCIF pdbx_poly_seq_scheme table with the mapping between
IHM and PMI numbering.