Remember the order we created the molecules in
Since we store molecules in a dict, they will get
built in random order, which will make RMF files
look a little odd (and make it difficult to compare
files, since RMF wants the hierarchy to be in the
same order each time). Address this by using an
OrderedDict instead, which remembers the order in
which molecules were created.