Fix mapping of PMI representation to ProtocolOutput
Rather than trying to extract the PMI representation
and pass it to ProtocolOutput at the point where it
was created, extract it later on when we search for
the highest resolution representation. This ensures
that we match what goes into the RMF file and gives
us the residues in the correct order.