Use ihm chemdescriptors for cross links
The python-ihm library now uses a ChemDescriptor
object to describe the chemistry of the linker
used in a cross linking experiment (as a SMILES or
INCHI string). Replace the existing linker type
(just a string) with a suitable object, creating a
dummy one if we can't figure out the linker type
from the file label.