Add support for DNA sequences
Replace the old 'is_nucleic' flag (which only supports RNA)
with a new 'alphabet' parameter. This allows any residue
alphabet to be used to map FASTA codes to residue types.
Mappings are provided for amino acids, RNA, and DNA
in the new IMP.pmi.alphabets module, but custom mappings
could also be added if desired.