Deprecate PMI functions to read/write mmCIF files
It is cleaner to use the python-ihm library directly
to read or write mmCIF files, so deprecate the PMI
functionality that simply duplicates this (passing a
file handle to the ProtocolOutput constructor, the
flush() method, and the read() function). Users should
use ProtocolOutput.finalize() and then ihm.dumper.write()
to output mmCIF/BinaryCIF, and get_dumpers()/get_handlers()
to write/read custom PMI information to/from mmCIF/BinaryCIF
files.