Make both atomic and non-atomic residues flexible
Previously, if a domain was assigned a PDB file but
not made rigid, we only added movers for the non-atomic
residues, i.e. those missing from the PDB file. See #251.
We should also add movers for the atomic residues so that
the domain is actually sampled flexibly, as requested by
the user. Make the associated warning message a little
clearer and suggest alternative approaches.