Only make things flexible if asked for
If no rigid bodies are specified for a domain, have
the default be to *not* move atomic residues, only the
missing regions. This should be a better fix for #251,
since 1) previously published systems (such as ECM29)
rely on this behavior and 2) it can be overridden if
desired by giving a PDB name of 'BEADS'.