Fix keep_chain_id when first component is not PDB
When building a component using keep_chain_id=True,
use the first PDB segment to assign the chain ID,
not the first segment of any kind (as if it is BEADS
it has a null chain ID). If there are no PDB segments,
emit a warning and use the default (sequential)
chain ID. Closes #252.