Trim non-represented residues from IHM sequence
PDB policy is for amino acid Entity sequences to be polymers (so
they should include any loops or other gaps in the midst of the
sequence) but for the termini to be trimmed of any not-modeled
residues. Add a function to, after the PMI system is built,
modify the Entity sequence to remove any parts that are not
included in any representation. This will change the seq_id numbering
if any N-terminal residues are removed. If this happens, set the
existing PMI-to-IHM offset accordingly, and renumber any
existing objects.